refactor: view 및 domain 패키지 생성
